requirements.txt
asgiref==3.9.1
boto3==1.40.14
certifi==2025.8.3
Django==5.2.5
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
Pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-decouple==3.8
whitenoise==6.9.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
dj-database-url==3.0.1
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0


# anyio==4.12.0
# asgiref==3.9.1
# backports.tarfile==1.2.0
# boto3==1.40.14
# botocore==1.40.14
# build==1.3.0
# CacheControl==0.14.4
# certifi==2025.8.3
# cffi==2.0.0
# charset-normalizer==3.4.3
# cleo==2.1.0
# colorama==0.4.6
# contourpy==1.3.3
# crashtest==0.4.1
# cryptography==46.0.3
# cycler==0.12.1
# defusedxml==0.7.1
# distlib==0.4.0
# dj-database-url==3.0.1
# Django==5.2.5
# django-cors-headers==4.7.0
# django-storages==1.14.6
# djangorestframework==3.16.1
# djangorestframework_simplejwt==5.5.1
# dulwich==0.24.10
# fastjsonschema==2.21.2
# filelock==3.20.0
# findpython==0.7.1
# fonttools==4.60.1
# gunicorn==23.0.0
# h11==0.16.0
# httpcore==1.0.9
# httpx==0.28.1
# idna==3.10
# importlib_metadata==8.7.0
# installer==0.7.0
# jaraco.classes==3.4.0
# jaraco.context==6.0.1
# jaraco.functools==4.3.0
# jmespath==1.0.1
# keyring==25.7.0
# kiwisolver==1.4.9
# matplotlib==3.7.2
# more-itertools==10.8.0
# msgpack==1.1.2
# numpy==2.3.4
# oauthlib==3.3.1
# packaging==25.0
# pandas==2.0.3
# pbs-installer==2025.12.5
# pillow==11.3.0
# pkginfo==1.12.1.2
# platformdirs==4.5.1
# plotly==5.15.0
# poetry==2.2.1
# poetry-core==2.2.1
# psycopg2-binary==2.9.10
# pycparser==2.23
# PyJWT==2.10.1
# pyparsing==3.0.9
# pyproject_hooks==1.2.0
# python-dateutil==2.9.0.post0
# python-decouple==3.8
# python-dotenv==1.1.1
# python3-openid==3.2.0
# pytz==2025.2
# pywin32-ctypes==0.2.3
# RapidFuzz==3.14.3
# requests==2.32.5
# requests-oauthlib==2.0.0
# requests-toolbelt==1.0.0
# s3transfer==0.13.1
# shellingham==1.5.4
# six==1.17.0
# social-auth-app-django==5.6.0
# social-auth-core==4.8.1
# sqlparse==0.5.3
# tenacity==9.1.2
# tomlkit==0.13.3
# trove-classifiers==2025.12.1.14
# typing_extensions==4.15.0
# tzdata==2025.2
# urllib3==2.5.0
# virtualenv==20.35.4
# whitenoise==6.9.0
# zipp==3.23.0
# zstandard==0.25.0
